+++ src/main/java/org/gautelis/durga/plugins/FieldFilter.java
@@ -21,8 +21,7 @@ import java.util.stream.Collectors;
 public final class FieldFilter implements Plugin {
 
     @Override
-    public byte[] execute(byte[] payload, String config) throws Exception {
-        String payloadStr = Plugin.toString(payload);
+    public String execute(String payload, String config) throws Exception {
         String keep = null;
         String drop = null;
         String flatten = null;
@@ -41,7 +40,7 @@ public final class FieldFilter implements Plugin {
                 }
             }
         }
-        return Plugin.toBytes(filter(payloadStr, keep, drop, flatten));
+        return filter(payload, keep, drop, flatten);
     }
 
     private FieldFilter() {