+++ src/main/java/org/gautelis/durga/plugins/RegexExtract.java
@@ -28,8 +28,7 @@ import java.util.regex.Pattern;
 public final class RegexExtract implements Plugin {
 
     @Override
-    public byte[] execute(byte[] payload, String config) throws Exception {
-        String payloadStr = Plugin.toString(payload);
+    public String execute(String payload, String config) throws Exception {
         String source = null;
         String pattern = null;
         String target = null;
@@ -51,7 +50,7 @@ public final class RegexExtract implements Plugin {
                 }
             }
         }
-        return Plugin.toBytes(extract(payloadStr, source, pattern, target, all));
+        return extract(payload, source, pattern, target, all);
     }
 
     private RegexExtract() {