+++ src/main/java/org/gautelis/durga/plugins/TimestampNormalize.java
@@ -43,8 +43,7 @@ public final class TimestampNormalize implements Plugin {
     private static final Logger LOG = LoggerFactory.getLogger(TimestampNormalize.class);
 
     @Override
-    public byte[] execute(byte[] payload, String config) throws Exception {
-        String payloadStr = Plugin.toString(payload);
+    public String execute(String payload, String config) throws Exception {
         String pluginName = "timestamp-normalize";
         Counter counter = Metrics.registry().counter("plugin.executions", "plugin", pluginName);
         Timer timer = Metrics.registry().timer("plugin.duration", "plugin", pluginName);
@@ -73,7 +72,7 @@ public final class TimestampNormalize implements Plugin {
                     }
                 }
             }
-            return Plugin.toBytes(normalize(payloadStr, fieldsList, from, to, zone, removeOnError));
+            return normalize(payload, fieldsList, from, to, zone, removeOnError);
         });
     }